requirements.txt
# Battery LLM Project Requirements

# Core Dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Vector Database
chromadb>=0.4.0
qdrant-client>=1.7.0

# Hugging Face
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0

# Document Loading
pypdf>=3.17.0
pytesseract>=0.3.10
pdfminer.six>=20221105
unstructured>=0.11.0

# Fine-tuning
trl>=0.7.0
deepspeed>=0.12.0

# MLOps
mlflow>=2.12.0
dvc>=3.0.0

# API & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
gradio>=4.0.0

# Evaluation
nltk>=3.8.1
rouge-score>=0.1.2
bleach>=6.1.0
scikit-learn>=1.4.0

# Monitoring
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tiktoken>=0.5.2
tenacity>=8.2.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
